Cleaning up

+++ src/Utilities/Utilities.cpp
@@ -349,8 +349,8 @@ void allocate_plan_2d(
 ,   const fftwf_r2r_kind kind
 ,   const unsigned nb
 ){
-    int            nb_table[2]   = {N, N};
-    int            nembed[2]     = {N, N};
+    int            nb_table[2]   = {(int)N, (int)N};
+    int            nembed[2]     = {(int)N, (int)N};
     fftwf_r2r_kind kind_table[2] = {kind, kind};
 
     float* vec = (float*) fftwf_malloc(N * N * nb * sizeof(float));
@@ -376,8 +376,8 @@ void allocate_plan_1d(
 ,   const fftwf_r2r_kind kind
 ,   const unsigned nb
 ){
-    int nb_table[] = {N};
-    int nembed[]   = {N};
+    int nb_table[] = {(int)N};
+    int nembed[]   = {(int)N};
     fftwf_r2r_kind kind_table[1] = {kind};
 
     float* vec = (float*) fftwf_malloc(N * nb * sizeof(float));
